options window cutted out into new py file
